pyproject.toml
[project]
name = "myside"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.115.12",
    "google-adk>=0.5.0",
    "googlesearch-python>=1.3.0",
    "litellm>=1.68.1",
    "selenium>=4.32.0",
]
